+++ release_notes.md
@@ -1,5 +1,16 @@
 # Release Notes - cul-folio-edge
 
+## [4.0] - 2026-01-16
+### Added
+- RSpec tests for full coverage of class methods (DACCESS-97)
+
+### Changed
+- Update code comments to replace 'Okapi' with more generic references to API gateway (today Kong, tomorrow ...?)
+- Non-authentication methods will raise an `AuthenticationError` if called with a nil token (DACCESS-807)
+
+### Fixed
+- Properly handle network-level errors in authentication (DACCESS-807)
+
 ## [3.3] - 2026-01-05
 ### Changed
 - Upcoming Sunflower Folio change, we are moving from Okapi to Kong. Updated the `cancel` method to support Kong's smaller payload (DACCESS-673)